chore: 프로필 박스 스타일시트 분리
