+++ client/src/pages/MainPage/styles.ts
@@ -14,56 +14,6 @@ export const profileListStyle = css({
   height: `calc(100% - 50px)`,
 });
 
-export const profileBoxStyle = css({
-  width: 'calc((100% / 3) - 40px)',
-  maxWidth: 600,
-  minWidth: 350,
-  maxHeight: 720,
-  display: 'grid',
-  gridTemplateRows: '10fr 1fr',
-  flexGrow: 1,
-  height: '95%',
-  padding: 20,
-  borderRadius: COMMON_SIZE.BORDER_RADIUS,
-  backgroundColor: COLORS.WHITE,
-  border: `${COMMON_SIZE.LINE_WIDTH}px solid ${COLORS.BOX_BORDER}`,
-  gap: 20,
-  textAlign: 'start',
-});
-
-export const textWrapperStyle = css({
-  display: 'flex',
-  flexDirection: 'column',
-});
-
-export const topTextStyle = css({
-  marginBottom: 5,
-
-  ' span': {
-    color: COLORS.TEXT_1,
-    fontWeight: '600',
-  },
-
-  ' span:first-of-type': {
-    marginRight: 10,
-  },
-});
-
-export const bottomTextStyle = css({
-  color: COLORS.TEXT_2,
-  opacity: 0.6,
-  fontWeight: '600',
-});
-
-export const profileBoxBottomStyle = css({
-  color: COLORS.TEXT_1,
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'space-between',
-  gap: 10,
-  fontWeight: 'bold',
-});
-
 export const emptyProfileBoxStyle = css({
   width: 'calc((100% / 3) - 40px)',
   maxWidth: 600,
@@ -73,22 +23,6 @@ export const emptyProfileBoxStyle = css({
   height: '95%',
   borderRadius: COMMON_SIZE.BORDER_RADIUS,
   backgroundColor: 'none',
-});
-
-export const favoriteButtonStyle = css({
-  borderRadius: '50%',
-  width: 40,
-  height: 40,
-  backgroundColor: 'white',
-  display: 'flex',
-  justifyItems: 'center',
-  alignItems: 'center',
-});
-
-export const favoriteIconStyle = css({
-  width: 30,
-  height: 30,
-  fill: COLORS.PRIMARY_2,
 });
 
 export const paginationStyle = css({
